chore: move adapter to new file, ref #4
