+++ code/ac/llama/Model.cpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 //
 #include "Model.hpp"
+#include "LoraAdapter.hpp"
 #include <llama.h>
 #include <astl/move.hpp>
 #include <stdexcept>
@@ -77,16 +78,6 @@ std::string Model::getChatTemplateId() const {
     }
 
     return std::string(tplBuf.get(), len);
-}
-
-LoraAdapter::LoraAdapter(Model& model, std::string path, float scale)
-    : m_adapter(llama_lora_adapter_init(model.lmodel(), path.c_str()), llama_lora_adapter_free)
-    , m_scale(scale)
-    , m_path(std::move(path))
-{
-    if (!m_adapter) {
-        throw std::runtime_error("Failed to initialize LORA adapter from " + m_path);
-    }
 }
 
 std::shared_ptr<llama_model> Model::ModelRegistry::loadModel(
