+++ code/ac/llama/Model.hpp
@@ -73,18 +73,4 @@ private:
     static ModelRegistry s_modelRegistry;
 };
 
-class AC_LLAMA_EXPORT LoraAdapter {
-public:
-    LoraAdapter(Model& model, std::string path, float scale = 1.0f);
-
-    llama_lora_adapter* adapter() const noexcept { return m_adapter.get(); }
-    float scale() const noexcept { return m_scale; }
-    std::string_view path() const noexcept { return m_path; }
-
-private:
-    astl::c_unique_ptr<llama_lora_adapter> m_adapter;
-    float m_scale;
-    std::string m_path;
-};
-
 } // namespace ac::llama
